telegram_bot/requirements.txt
python-telegram-bot==20.7
ctypes-utils==0.1.0
python-dotenv==1.0.0 